% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{check_upload_status}
\alias{check_upload_status}
\title{Check the status of a Mapbox upload}
\usage{
check_upload_status(username, id, access_token = NULL)
}
\arguments{
\item{username}{Your account's username}

\item{id}{The upload ID}

\item{access_token}{Your Mapbox access token}
}
\description{
Check the status of a Mapbox upload
}
