% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{feature_options}
\alias{feature_options}
\title{Specify feature options for an MTS recipe layer}
\usage{
feature_options(
  id = NULL,
  bbox = NULL,
  attributes = list(zoom_element = NULL, set = NULL, allowed_output = NULL),
  filter = NULL,
  simplification = NULL
)
}
\arguments{
\item{id}{A column representing the feature ID. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#id-expression}.}

\item{bbox}{A bounding box within which rendered features will be clipped. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#bounding-box}.}

\item{attributes}{A named list of attribute transformations. \code{zoom_element} specifies how an attribute should be made available at different zoom levels; \code{set} allows you to calculate new attributes from existing attributes when processing the tiles; and \code{allowed_output} specifies which columns should be carried through to the output tiles.  See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-attributes}.}

\item{filter}{An expression that determines how features in the tileset should be filtered. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-filters} for information on how to specify the filter.}

\item{simplification}{Rules for feature simplification.  See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-simplification} for more information on how to specify this.}
}
\value{
A list of feature options, likely to be used in \code{recipe_layer()}.
}
\description{
Specify feature options for an MTS recipe layer
}
