% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{geocoder_as_sf}
\alias{geocoder_as_sf}
\title{Convert the result of a \code{mapboxGeocoderInput()} geocoded location to an sf object}
\usage{
geocoder_as_sf(input)
}
\arguments{
\item{input}{The name of the Shiny input using \code{mapboxGeocoderInput()}, likely something like \code{input$geocode}}
}
\value{
An sf object that can be used downstream in your Shiny applications
}
\description{
Convert the result of a \code{mapboxGeocoderInput()} geocoded location to an sf object
}
