% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_geocoder_dependencies}
\alias{get_geocoder_dependencies}
\title{Include Mapbox Geocoder dependencies}
\usage{
get_geocoder_dependencies()
}
\description{
This function loads necessary JavaScript and CSS for the Mapbox Geocoder.
}
\keyword{internal}
