% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{get_style}
\alias{get_style}
\title{Get information about a style from your Mapbox account}
\usage{
get_style(style_id, username, style_url = NULL, access_token = NULL)
}
\arguments{
\item{style_id}{A style ID}

\item{username}{A Mapbox username}

\item{style_url}{Mapbox style URL}

\item{access_token}{A Mapbox public or secret access token; set with \code{mb_access_token()}}
}
\value{
A list of information about your selected style.
}
\description{
Get information about a style from your Mapbox account
}
