% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{list_styles}
\alias{list_styles}
\title{List styles in your Mapbox account}
\usage{
list_styles(username, access_token = NULL)
}
\arguments{
\item{username}{Your Mapbox username}

\item{access_token}{Your Mapbox public or secret access token; set with \code{mb_access_token()}}
}
\value{
A data frame of information about styles in your Mapbox account
}
\description{
List styles in your Mapbox account
}
