% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{mb_directions}
\alias{mb_directions}
\title{Make a request to the Mapbox Directions API}
\usage{
mb_directions(
  input_data = NULL,
  origin = NULL,
  destination = NULL,
  profile = "driving",
  output = "sf",
  alternatives = NULL,
  annotations = NULL,
  bearings = NULL,
  continue_straight = NULL,
  exclude = NULL,
  geometries = NULL,
  overview = "simplified",
  radiuses = NULL,
  approaches = NULL,
  steps = NULL,
  banner_instructions = NULL,
  language = NULL,
  roundabout_exits = NULL,
  voice_instructions = NULL,
  voice_units = NULL,
  waypoint_names = NULL,
  waypoint_targets = NULL,
  waypoints = NULL,
  walking_speed = NULL,
  walkway_bias = NULL,
  alley_bias = NULL,
  access_token = NULL
)
}
\arguments{
\item{input_data}{An input dataset of class \code{"sf"}, or a list of coordinate pairs for format \code{c(longitude, latitude)}. Cannot be used with an origin/destination pair.}

\item{origin}{An address or coordinate pair that represents the origin of your requested route. Cannot be used with \code{input_data}.}

\item{destination}{An address or coordinate pair that represents the destination of your requested route.}

\item{profile}{One of "driving" (the default), "driving-traffic", "walking", or "cycling".}

\item{output}{One of "sf" (the default), which returns an sf LINESTRING representing the route geometry, or "full", which returns the full request from the Directions API as a list.}

\item{alternatives}{Whether or not to return alternative routes with your request.}

\item{annotations}{Must be used with overview = FULL.}

\item{bearings}{bearings}

\item{continue_straight}{continue_straight}

\item{exclude}{exclude}

\item{geometries}{geometries}

\item{overview}{overview}

\item{radiuses}{radiuses}

\item{approaches}{approaches}

\item{steps}{steps}

\item{banner_instructions}{banner_instructions}

\item{language}{language}

\item{roundabout_exits}{roundabout_exits}

\item{voice_instructions}{voice_instructions}

\item{voice_units}{voice_units}

\item{waypoint_names}{waypoint_names}

\item{waypoint_targets}{waypoint_targets}

\item{waypoints}{waypoints}

\item{walking_speed}{walking_speed}

\item{walkway_bias}{walkway_bias}

\item{alley_bias}{alley_bias}

\item{access_token}{acces_token}
}
\value{
A list or sf object
}
\description{
Make a request to the Mapbox Directions API
}
