% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_get_recipe}
\alias{mts_get_recipe}
\title{Retrieve the recipe for an MTS tileset in your Mapbox account}
\usage{
mts_get_recipe(tileset_name, username, access_token = NULL)
}
\arguments{
\item{tileset_name}{The tileset name for which you'd like to retrieve a recipe}

\item{username}{Your Mapbox username}

\item{access_token}{Your Mapbox access token with secret scope}
}
\value{
The recipe for your tileset as an R list
}
\description{
Retrieve the recipe for an MTS tileset in your Mapbox account
}
