% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_make_recipe}
\alias{mts_make_recipe}
\title{Make a recipe list for use with the Mapbox Tiling Service}
\usage{
mts_make_recipe(...)
}
\arguments{
\item{...}{One or more named lists that represent layers in the Mapbox Tiling Service recipe specification (\url{https://docs.mapbox.com/mapbox-tiling-service/reference/#layer-example}).  If multiple layers are included, a multi-layer recipe will be prepared that can produce tilesets with multiple sources.}
}
\value{
An R list representing an MTS recipe to be used to create a tileset.
}
\description{
Make a recipe list for use with the Mapbox Tiling Service
}
