% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_publish_tileset}
\alias{mts_publish_tileset}
\title{Publish a tileset with Mapbox Tiling Service}
\usage{
mts_publish_tileset(tileset_name, username, access_token = NULL)
}
\arguments{
\item{tileset_name}{The name of the tileset (as supplied to \code{mts_create_tileset()})}

\item{username}{Your Mapbox username}

\item{access_token}{Your Mapbox access token}
}
\value{
The API response
}
\description{
Publish a tileset with Mapbox Tiling Service
}
