% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_make_recipe}
\alias{mts_make_recipe}
\alias{recipe_layer}
\alias{feature_options}
\alias{tile_options}
\title{Make a recipe list for use with the Mapbox Tiling Service}
\usage{
mts_make_recipe(...)

recipe_layer(
  source,
  minzoom,
  maxzoom,
  features = feature_options(),
  tiles = tile_options()
)

feature_options(
  id = NULL,
  bbox = NULL,
  attributes = list(zoom_element = NULL, set = NULL, allowed_output = NULL),
  filter = NULL,
  simplification = NULL
)

tile_options(
  bbox = NULL,
  extent = NULL,
  buffer_size = NULL,
  limit = NULL,
  union = list(where = NULL, group_by = NULL, aggregate = NULL, maintain_direction =
    NULL, simplification = NULL),
  filter = NULL,
  attributes = NULL,
  order = NULL,
  remove_filled = NULL,
  id = NULL,
  layer_size = NULL
)
}
\arguments{
\item{...}{One or more named lists that represent layers in the Mapbox Tiling Service recipe specification (\url{https://docs.mapbox.com/mapbox-tiling-service/reference/#layer-example}).  These lists can be prepared with the helper function \code{recipe_layer()}, or prepared by hand if the user prefers. If multiple layers are included, a multi-layer recipe will be prepared that can produce tilesets with multiple sources.}

\item{source}{The tileset source ID. This is returned by \code{mts_create_source()} or can be retrieved from your Mapbox account with \code{mts_list_sources()}.}

\item{minzoom}{The minimum zoom level at which a layer can be viewed.}

\item{maxzoom}{The maximum zoom level at which a layer is rendered; the layer will still be visible past the maximum zoom level due to overzooming.}

\item{features}{A list of feature options, possibly generated with \code{feature_options()}.}

\item{tiles}{A list of tile options, possibly generated with \code{tile_options()}}

\item{id}{A column representing the feature ID. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#id-expression}.}

\item{bbox}{A bounding box within which rendered features will be clipped. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#bounding-box}.}

\item{attributes}{A named list of attribute transformations. \code{zoom_element} specifies how an attribute should be made available at different zoom levels; \code{set} allows you to calculate new attributes from existing attributes when processing the tiles; and \code{allowed_output} specifies which columns should be carried through to the output tiles.  See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-attributes}.}

\item{filter}{An expression that determines how features in the tileset should be filtered. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-filters} for information on how to specify the filter.}

\item{simplification}{Rules for feature simplification.  See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#feature-simplification} for more information on how to specify this.}

\item{bbox, extent, buffer_size, limit, union, filter, attributes, order, remove_filled, id, layer_size}{Tile options in the MTS recipe. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#tile-configuration} for more information on the available options.}
}
\value{
An R list representing an MTS recipe to be used to create a tileset.

A recipe layer list to be used in \code{mts_make_recipe()}.

A list of feature options, likely to be used in \code{recipe_layer()}.

A list of tile options, likely to be used in \code{recipe_layer}.
}
\description{
Make a recipe list for use with the Mapbox Tiling Service

Build a recipe layer for use in an MTS recipe

Specify feature options for an MTS recipe layer

Specify tile options for an MTS recipe layer
}
