% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_update_recipe}
\alias{mts_update_recipe}
\title{Update a tileset's MTS recipe}
\usage{
mts_update_recipe(tileset_name, username, recipe, access_token = NULL)
}
\arguments{
\item{tileset_name}{The name of your Mapbox tileset}

\item{username}{Your Mapbox username}

\item{recipe}{The new recipe for your tileset, likely created with \code{mts_make_recipe()}.}

\item{access_token}{Your Mapbox access token}
}
\value{
If the update is successful, the function will print a message informing you of its success.  Otherwise, a list of responses from the API will be returned letting you know why the request was invalid.
}
\description{
Update a tileset's MTS recipe
}
