% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_validate_recipe}
\alias{mts_validate_recipe}
\title{Validate a Mapbox Tiling Service recipe}
\usage{
mts_validate_recipe(recipe, access_token = NULL)
}
\arguments{
\item{recipe}{A recipe list, created with \code{mts_make_recipe()}}

\item{access_token}{Your Mapbox access token.}
}
\value{
A response from the API indicating whether the MTS recipe is valid or not. If the recipe is invalid, the API response will tell you the reason why.
}
\description{
Validate a Mapbox Tiling Service recipe
}
