% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{static_mapbox}
\alias{static_mapbox}
\title{Return a static Mapbox map from a specified style}
\usage{
static_mapbox(
  style_id,
  username,
  overlay_sf = NULL,
  overlay_markers = NULL,
  longitude = NULL,
  latitude = NULL,
  zoom = NULL,
  width = 600,
  height = 600,
  bearing = NULL,
  pitch = NULL,
  double_scale = FALSE,
  attribution = TRUE,
  logo = TRUE,
  before_layer = NULL,
  addlayer = NULL,
  setfilter = NULL,
  layer_id = NULL,
  access_token = NULL
)
}
\arguments{
\item{style_id}{The style ID}

\item{username}{Your Mapbox username}

\item{overlay_sf}{The overlay sf object, if applicable}

\item{overlay_markers}{The overlay markers, if applicable}

\item{longitude}{The longitude of the map center}

\item{latitude}{The latitude of the map center}

\item{zoom}{The map zoom}

\item{width}{The map width; defaults to 600px}

\item{height}{The map height; defaults to 600px}

\item{bearing}{The map bearing}

\item{pitch}{The map pitch}

\item{double_scale}{Whether or not to return a 2x map}

\item{attribution}{Not yet implemented}

\item{logo}{Not yet implemented}

\item{before_layer}{Not yet implemented}

\item{addlayer}{Not yet implemented}

\item{setfilter}{Not yet implemented}

\item{layer_id}{Not yet implemented}

\item{access_token}{Your Mapbox access token}
}
\value{
right now, an http request
}
\description{
Return a static Mapbox map from a specified style
}
