% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{tile_options}
\alias{tile_options}
\title{Specify tile options for an MTS recipe layer}
\usage{
tile_options(
  bbox = NULL,
  extent = NULL,
  buffer_size = NULL,
  limit = NULL,
  union = list(where = NULL, group_by = NULL, aggregate = NULL, maintain_direction =
    NULL, simplification = NULL),
  filter = NULL,
  attributes = NULL,
  order = NULL,
  remove_filled = NULL,
  id = NULL,
  layer_size = NULL
)
}
\arguments{
\item{bbox, extent, buffer_size, limit, union, filter, attributes, order, remove_filled, id, layer_size}{Tile options in the MTS recipe. See \url{https://docs.mapbox.com/mapbox-tiling-service/reference/#tile-configuration} for more information on the available options.}
}
\value{
A list of tile options, likely to be used in \code{recipe_layer}.
}
\description{
Specify tile options for an MTS recipe layer
}
