% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{upload_tiles}
\alias{upload_tiles}
\title{Upload dataset to your Mapbox account}
\usage{
upload_tiles(input, username, access_token = NULL, tileset_name = NULL)
}
\arguments{
\item{input}{The path to the dataset to upload}

\item{username}{Your Mapbox username}

\item{access_token}{Your Mapbox password}

\item{tileset_name}{The name of the tileset}
}
\description{
Upload dataset to your Mapbox account
}
