% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mb_access_token}
\alias{mb_access_token}
\title{Install a Mapbox access token in your .Renviron for repeated use}
\usage{
mb_access_token(token, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{token}{The Mapbox access token}

\item{overwrite}{Whether or not to overwrite an existing Mapbox access token.  Defaults to FALSE.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\description{
Install a Mapbox access token in your .Renviron for repeated use
}
\examples{
\dontrun{
my_token <- "..." # The token generated from your Mapbox account
mb_access_token(my_token, install = TRUE)
Sys.getenv("MAPBOX_ACCESS_TOKEN")
}
}
