% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{mb_geocode}
\alias{mb_geocode}
\title{Geocode an address using the Mapbox Geocoding API (skeleton at the moment)}
\usage{
mb_geocode(query, limit = 1, sf = FALSE, access_token = NULL)
}
\arguments{
\item{query}{The address query, formatted as a character string.}

\item{limit}{How many results to return; defaults to 1}

\item{sf}{If TRUE, returns the query result as an sf object.  Defaults to FALSE.}

\item{access_token}{The Mapbox access token (required); can be set with \code{mb_access_token}.}
}
\description{
Geocode an address using the Mapbox Geocoding API (skeleton at the moment)
}
