% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{mb_geocode}
\alias{mb_geocode}
\title{Geocode an address using the Mapbox Geocoding API (skeleton at the moment)}
\usage{
mb_geocode(query, limit = 1, sf = FALSE, access_token = NULL)
}
\arguments{
\item{access_token}{}
}
\description{
Geocode an address using the Mapbox Geocoding API (skeleton at the moment)
}
