% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{mb_isochrone}
\alias{mb_isochrone}
\title{Generate an isochrone using the Mapbox API}
\usage{
mb_isochrone(location, profile = "driving", contours = c(5, 10, 15),
  access_token = NULL, denoise = 1, geometry = "polygons",
  output = "sf")
}
\arguments{
\item{location}{A vector of form \code{c(longitude, latitude)} or an address that can be geocoded as a character string.}

\item{profile}{One of "driving", "walking", or "cycling".  "driving" is the default.}

\item{contours}{A vector of isochrone contours, specified in minutes.  4 is the maximum; defaults to \code{c(5, 10, 15)}.}

\item{access_token}{A valid Mapbox access token.}

\item{denoise}{XXX}

\item{polygons}{YYY}
}
\value{
An sf object representing the isochrone(s) around the location.
}
\description{
Generate an isochrone using the Mapbox API
}
