% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{mb_matrix}
\alias{mb_matrix}
\title{Retrieve a matrix of travel times from the Mapbox Directions API}
\usage{
mb_matrix(input_data, sources = "all", destinations = "all",
  mode = "driving", fallback_speed = 44, access_token = NULL,
  duration_output = "minutes")
}
\arguments{
\item{input_data}{The input coordinates of your request.  Acceptable inputs include a vector of
coordinate pairs in "long, lat" format or an sf object.
For sf linestrings or polygons, the distance between centroids will be taken.}

\item{sources}{The positional indices of the source coordinates.  Defaults to "all".}

\item{destinations}{The positional indices of the destination coordinates.  Defaults to "all".}

\item{mode}{One of "driving" (the default), "driving-traffic", "walking", or "cycling".}

\item{access_token}{A Mapbox access token (required)}
}
\value{
An R matrix of source-destination travel times.
}
\description{
Retrieve a matrix of travel times from the Mapbox Directions API
}
