% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_create_tileset}
\alias{mts_create_tileset}
\title{Create a tileset with the Mapbox Tiling Service API}
\usage{
mts_create_tileset(
  tileset_name,
  username,
  recipe,
  request_name = tileset_name,
  access_token = NULL
)
}
\arguments{
\item{tileset_name}{The name of the MTS tileset you intend to create}

\item{username}{Your Mapbox username}

\item{recipe}{An MTS recipe, created with \code{mts_make_recipe()}}

\item{request_name}{The name of the request; defaults to the tileset name}

\item{access_token}{Your Mapbox access token}
}
\value{
The response from the API, formatted as an R list.
}
\description{
After you've uploaded your spatial data to your Mapbox account with \code{mts_create_source} and prepared a valid recipe with \code{mts_make_recipe()}, you can use your source and recipe to create a vector tileset.  This tileset will be hosted at your Mapbox account.  Once created successfully, you will need to publish the tileset using \code{mts_publish_tileset} to use it in Mapbox Studio, Mapbox GL JS, or an R package that can read Mapbox tilesets.
}
