% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_create_tileset}
\alias{mts_create_tileset}
\title{Create a tileset with the Mapbox Tiling Service API}
\usage{
mts_create_tileset(
  recipe,
  tileset_name,
  username,
  request_name = tileset_name,
  access_token = NULL
)
}
\arguments{
\item{recipe}{An MTS recipe, created with \code{mts_make_recipe()}}

\item{tileset_name}{The name of the MTS tileset you intend to create}

\item{username}{Your Mapbox username}

\item{request_name}{The name of the request; defaults to the tileset name}

\item{access_token}{Your Mapbox access token}
}
\value{
The API response
}
\description{
Create a tileset with the Mapbox Tiling Service API
}
