% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_publish_tileset}
\alias{mts_publish_tileset}
\title{Publish a tileset with Mapbox Tiling Service}
\usage{
mts_publish_tileset(tileset_name, username, access_token = NULL)
}
\arguments{
\item{tileset_name}{The name of the tileset (as supplied to \code{mts_create_tileset()})}

\item{username}{Your Mapbox username}

\item{access_token}{Your Mapbox access token}
}
\value{
The response from the Mapbox Tiling Service API, formatted as an R list.
}
\description{
\code{mts_publish_tileset()} publishes an existing vector tileset at your Mapbox account, allowing you to use the vector tiles in your projects.  The tileset name will be the same name you specified in \code{mts_create_tileset()}.
}
\details{
The published tileset will conform to rules specified in its recipe.  If you want to change the recipe for a tileset, use \code{mts_update_recipe()} then re-publish the tileset with a call to \code{mts_publish_tileset()} once more.
}
