% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts.R
\name{mts_validate_recipe}
\alias{mts_validate_recipe}
\title{Validate a Mapbox Tiling Service recipe}
\usage{
mts_validate_recipe(recipe, access_token = NULL)
}
\arguments{
\item{recipe}{A recipe list, created with \code{mts_make_recipe()}}

\item{access_token}{Your Mapbox access token.}
}
\value{
A response from the API indicating whether the MTS recipe is valid or not. If the recipe is valid, returns \code{TRUE}, allowing you to use the output of this function for error handling pipelines.  If the recipe is invalid, the function returns \code{FALSE} and prints the API response telling you why the recipe is invalid.
}
\description{
Validate a Mapbox Tiling Service recipe
}
