% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tippecanoe.R
\name{tippecanoe}
\alias{tippecanoe}
\title{Generate an .mbtiles file with tippecanoe}
\usage{
tippecanoe(input, output, options = NULL)
}
\arguments{
\item{input}{The dataset from which to generate vector tiles.  Can be an sf object or GeoJSON file on disk.}

\item{output}{The name of the output .mbtiles file (with .mbtiles extension).  Will be saved in the current working directory.}

\item{options}{A character vector of options to be passed to the tippecanoe program.}
}
\description{
Generate an .mbtiles file with tippecanoe
}
