% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tippecanoe.R
\name{tippecanoe}
\alias{tippecanoe}
\title{Generate an .mbtiles file with tippecanoe}
\usage{
tippecanoe(
  input,
  output,
  min_zoom = NULL,
  max_zoom = NULL,
  drop_rate = NULL,
  overwrite = TRUE,
  other_options = NULL,
  keep_file = FALSE
)
}
\arguments{
\item{input}{The dataset from which to generate vector tiles.  Can be an sf object or GeoJSON file on disk.}

\item{output}{The name of the output .mbtiles file (with .mbtiles extension).  Will be saved in the current working directory.}

\item{min_zoom}{min_zoom}

\item{max_zoom}{max_zoom}

\item{drop_rate}{drop_rate}

\item{overwrite}{overwrite}

\item{other_options}{A character vector of options to be passed to the tippecanoe program.}

\item{keep_file}{Whether nor not to keep the temporary CSV or GeoJSON file used to generate the tiles.  Defaults to \code{FALSE}.}
}
\description{
Generate an .mbtiles file with tippecanoe
}
